Make logout inline
Ref. #16
