+++ skygear_example/src/main/AndroidManifest.xml
@@ -20,7 +20,6 @@
         </activity>
         <activity android:name=".SignupActivity" android:screenOrientation="portrait" />
         <activity android:name=".LoginActivity" android:screenOrientation="portrait" />
-        <activity android:name=".LogoutActivity" android:screenOrientation="portrait" />
         <activity android:name=".RecordCreateActivity" android:screenOrientation="portrait" />
         <activity android:name=".RecordQueryActivity" android:screenOrientation="portrait" />
     </application>
