+++ skygear_example/src/main/java/io/skygear/skygear_example/MainActivity.java
@@ -1,13 +1,17 @@
 package io.skygear.skygear_example;
 
+import android.app.ProgressDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.TextView;
 
 import io.skygear.skygear.Configuration;
 import io.skygear.skygear.Container;
+import io.skygear.skygear.LogoutResponseHandler;
 import io.skygear.skygear.User;
 
 public class MainActivity extends AppCompatActivity {
@@ -41,6 +45,10 @@ public class MainActivity extends AppCompatActivity {
         this.endpointDisplay.setText(skygearConfig.endpoint);
         this.apiKeyDisplay.setText(skygearConfig.apiKey);
 
+        this.updateUserInfoDisplay();
+    }
+
+    private void updateUserInfoDisplay() {
         User currentUser = this.skygear.getCurrentUser();
         if (currentUser != null) {
             this.accessTokenDisplay.setText(currentUser.accessToken);
@@ -51,7 +59,55 @@ public class MainActivity extends AppCompatActivity {
             this.userIdDisplay.setText(R.string.undefined);
             this.usernameDisplay.setText(R.string.undefined);
         }
+    }
 
+    public void doLogout(View view) {
+        new AlertDialog.Builder(this)
+                .setTitle("Confirm")
+                .setMessage("Are you sure to logout ?")
+                .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        MainActivity.this.doLogoutWithConfirmation();
+                    }
+                })
+                .setNeutralButton("No", null)
+                .show();
+    }
+
+    private void doLogoutWithConfirmation() {
+        final ProgressDialog loading = new ProgressDialog(this);
+        loading.setTitle("Loading");
+        loading.setMessage("Logging out...");
+        loading.show();
+
+        final AlertDialog successDialog = new AlertDialog.Builder(this)
+                .setTitle("Logout success")
+                .setMessage("You have logged out.")
+                .create();
+
+        final AlertDialog failDialog = new AlertDialog.Builder(this)
+                .setTitle("Logout failed")
+                .setMessage("")
+                .create();
+
+        this.skygear.logout(new LogoutResponseHandler() {
+            @Override
+            public void onLogoutSuccess() {
+                loading.dismiss();
+                successDialog.show();
+
+                MainActivity.this.updateUserInfoDisplay();
+            }
+
+            @Override
+            public void onLogoutFail(String reason) {
+                loading.dismiss();
+
+                failDialog.setMessage("Fail with reason: \n" + reason);
+                failDialog.show();
+            }
+        });
     }
 
     public void goSignup(View v) {
@@ -60,10 +116,6 @@ public class MainActivity extends AppCompatActivity {
 
     public void goLogin(View v) {
         startActivity(new Intent(this, LoginActivity.class));
-    }
-
-    public void goLogout(View view) {
-        startActivity(new Intent(this, LogoutActivity.class));
     }
 
     public void goRecordCreateDelete(View view) {
